begin translation
started the native rust port of the dev branch of C++ impl.
i'm planning on changing a lot of the impl details to be more rustic.

i've also decided to make some more restrictions on user inputs, based on type-safety in rust.
for instance, expression operations involving mixed float/int will no longer convert the int to float - it's simply an error.
there are a few other related changes for other expr ops.

this commit includes fully functioning symbol table, expr evaluation, and binary sets (renamed from C++ impl's binarly literal collection).
this also includes a fully functional cross-platform binary serializer, which is imlemented for all types that will need it down the road.
it's not as compact as the custom logic used in the C++ impl, but it's more modular and ergonomic - the size difference is not really a concern for me at this point.
the CPURegister type is also implemented, but it's kind of alone for the time being - i intend to get the assembler finished first.

most everything has some form of unit test, esp. the collection types.

i found a minor bug in the binary set logic used by the C++ impl - it doesn't cause it to be less efficient from an endpoint perspective, but sometimes makes a duplicate slice entry internally.