beginning AsmArgs
beginning the AsmArgs translation.
i had intended to just impl Expr extraction, but that's turning out to be way more involved than initially intended.
for instance, it refers to several non-trivial helper functions which had to be written.
additionally, the cpp impl logic uses pointers, which is impossible in rust without using unsafe (which i've explicitly forbid).
so now i need to impl a different approach using the Shunting-yard algorithm.
it's not so complicated, but will be more time consuming due to not having the code to translate from.

this commit adds a fully functioning extract_string() helper, which is needed by extract_expr().
also more or less finalized the AsmError struct to have exactly what it needs and nothing more.
this still includes significantly more info than the cpp/cs impls give.
included is an enum denoting the error (as opposed to the cpp/cs versions giving a string version of the same).

i've also decided to no longer support back tick (`) strings.
these used to be for enabling character escapes, which (') and (") did not do (they were raw strings).
now we only support ' and ", which both allow escapes.
it's possible i might re-add back ticks as raw strings - i just want escapes to be the default behavior.