more work on AsmArgs
i've realized i want to change the ways exprs are parsed to include string literals.
this would make parsing higher-level structures simpler and lets us restrict how strings are used in an expr.
currently strings in an expr are just converted to 8-byte (typically 8 ascii char) and stored as u64.
they then operate identically to u64, which is perhaps not ideal (e.g. 'abc' + 'd' != 'abcd' but is valid).
this was originally done because c#/cpp impls for everything else were already done and strings were more of an extension.

this commit stores work prior to strings being expr.