expr refactor
made some changes to expr in preparation for adding expr string type.
changed expr::Value to no longer be Copy (but is still Clone).
refactored expr to return contents by reference.
required introducing a smart pointer type to hold an owning handle.
the handle has to have a (mutable) borrow of the refcell contents, meaning we cannot allow it to escape the crate.
consequently, eval() is no longer public, a worthy sacrifice.