finished core expr parsing
finished the core behaviors of expr parsing with the shunting-yard algorithm.
i decided to abandon the ?: conditional operator in favor of the $if functional macro.
this is both because the associativity of ?: is somewhat strange and $if is vastly easier to implement.

additionally, i decided to drop case-insensitive macros.
aesthetically, this is because e.g. $PtR looks dumb but would be valid, and forcing all lowercase is more regulated.
I also plan on adding user-defined functional macros at some point, and those would be case sensitive, so they might as well all be.