fpu/f80 and more asm parsing
added the FPU type, along with special types for each of its registers which expose their fields/flags.
FPU uses the rug::Float type set to 64-bit precision to accurately emulate x87 80-bit floats.

added the f80 module with the key type F80, representing an 80-bit floating point value.
this will not be used for doing arithmetic directly - that's handled as rug::Float.
F80 is simply an array of 10 bytes which has functions to convert back and forth from it to Float.
this will be used to support fstenv and tbyte float encodings.
this was one of the major benefits of porting everything over to rust - the preponderance of high-quality libraries made this possible.
none of the other impls support this, and adding it would be a pain, so new csx64 standard-conforming impls will likely just be rust wrappers.

added some more stuff for asm parsing.
the Argument enum type will be used for all args, be it cpu/fpu/vpu reg, address, or imm.
got a good amount of the address parsing done, but the hard part is of course implementing the reg mult algebra.
that should be the final pitfall for the assembler, though.