linker added
added the `link()` function to link object files into executables.
the translation wasn't so bad, but a few things had to change in order to be ergonomic.
for one, we no longer do the weird top level binary literal merging.
looking at it now, i'm not even sure it's safe to do in the general case, yet it apparently works for everything i'd tested on.
new version just adds to the set as per usual, which is slower (but still linear) but definitely safe.

changed a few other things to have better error messages.
for instance, SymbolTable is now generic on a tag type `T`, which is used by assembler/linker to keep track of decl line numbers.
to avoid monomorphic explosion, added the `SymbolTableCore` trait which implements all the core non-tag operations.
we then take a `&dyn SymbolTableCore` instead of `&SymbolTable` for expr eval contexts.

added linker tests file, but it's basically nothing as of now - just committing since it's in a theoretically working state.
onward to link tests and the exec engine!