init/exec and impl mov/movcc
added init/exec logic and implemented mov and movcc.
also implemented syscall, but currently only has exit.

added unit tests for exec stuff.
the exec engine appears to be working just fine, but this caught a bunch of bugs in the assembler/linker.
these have been corrected, but there are likely more.
they have to do with expr internals, specifically with the `take_or_*` logic, which if not handled properly can be disasterous.
added a new unique illegal value category to help make this not silently succeed, but it has the effect of panicing if encountered.
this is better than the alternative, but not by much.
i'm planning on reworking the whole expr ownership thing to solve the problem more gracefully.