major expr refactor
refactored expr to get rid of the dangerous `take_or_*` family of methods.
`ValueRef` now is purely an immutable smart pointer to a value.
To get the same utility as `take_or_*`, i've added `ValueCow` which is similar to `Cow`.
the important usage difference is that the function which yields `ValueCow` consumes the expr.
this offsets the safety considerations to the compiler, and is therefore always sound.

also, this new logic fixes a bug where `a + a` would panic due to mutably borrowing `a` twice.
added test cases for this and the double-cyclic form `a = b + b`, `b = a + a`.
to facilitate this, we use a cobination of mutable and immutable borrowing and have to maintain a separate `visited` set of identifiers.

removed signed/unsigned values - now a single monolithic integer type.
this is implemented with rug::Integer.
the intention is to make compile-time literals be completely abstracted from physical representation, which was not the case before due to using finite sized types and being explicitly signed or unsigned.
i've set a hard cap for the maximum number of significant bits in an int as a safety feature.
exceeding this limit results in an eval error.
these limits aren't enforced for many instructions right now (just shl and shr), but will be expanded later.

currently, this uses the (unstable) move_ref_pattern for conciceness and readability, which means the nightly compiler is needed for now.
however, this pattern was scheduled for release in rust 1.49, so i figure it's ok.
we'll continue work in nightly until it's stabilized.