added lea
added lea instruction and migrated over the c++ impl tests for it.
updated address parsing error handling to give better error messages.
before, failure to parse address would just make it parse as expr and give error for that.
now, at certain points in address parsing we return a different type denoting that it failed but should be treated as an address.