added jmp/jcc/loopcc/call/ret
added the above instructions.
required implementing the "immrm" encoding from the C++ impl, which is now renamed to "value" to denote that it's any value, but not necessarily an lvalue like "unary" is.
added tests for all the new instructions as usual.
this caught a bug in the assembler where we were skipping the last nonlocal label update code if no instruction or times = 0 due to a continue statement.
this is fixed by moving the "non-continue" code inside the match statement and doing times iter checks before loop logic.

refactored the conditional ops flag logic to be its own function.
different conditional ops have different conditions they support, but that's more-so a problem for the assembler.
the emulator can union them together and blindly support all of them for everything.