added inc/dec/neg/not
added the above instructions.
the c++ impl for inc/dec actually had a bit of non-standard logic.
it was setting CF based on unsiged overflow/underflow, but apparently the x86 instructions do not do this.
this is slightly annoying, but i've decided to side with the standard instructions.

changed the core NULL constant to be an integer instead of pointer.
this is in lieu of relocatable code generation, which would need to differentiate offset address 0 from physical address 0.
so far, relocatability is planned to be an O(n) operation of applying offsets to all stored pointers.
this is not ideal, but it's much simpler than redoing everything in terms of position independent code.

fixed an asm error message bug where late expr evaluations were using the current line number instead of the line that defined it.