added mul/imul
added the mul and imul instructions.

added the ternary op format, which despite the name is in fact different from the ternary format in the c++ impl.
this was due to a misconception about how mulx works (a bug in the c++ impl), which has 2 destinations, not 2 sources.
the new format supports one or two sources AND/OR one or two destinations, depending on context.
additionally, it's been simplified to merge the old RR_RM format from the c++ impl (which will not be present in this impl).
ternary is simply a register prefix, followed by a binary format op.
this makes it more featureful while also reducing code dupe and complexity.

standard x86 has unary mul and unary/binary/ternary imul.
i considered this to be annoying, so i gave mul binary/ternary forms as well.
aside from computing CF/OF differently, these are the same as their imul equivalents.