io syscalls + fs module
assemble() now takes a different type: Predefines, which is a thin wrapper for SymbolTable.
the reason for this is that SymbolTable::default() should be an empty symbol table,
while Predefines::default() should be the set of all default predefined symbols (e.g. syscall codes).
additionally this prevents users from falsely providing line numbers for predefine origins (should all be 0).

added the fs module, which notably has the FileHandle trait.
these are held as `Vec<Option<Arc<Mutex<dyn FileHandle>>>>` by the emulator with the index being the file descriptor.
the `Arc<Mutex<T>>` part is so that we can dupe file descriptors and share across threads, which the C++ and C# impls don't allow.

added MemoryFile, which is currently the only implementor of FileHandle.
this just stores all its content in memory, which makes it easy to do unit tests.

renamed $ptr macro to $bin - will be more consistent with type casting macros in the future.

added FusedIterator to impl bound for SymbolTable.