added BTx family of instructions
added the BTx family of instructions (used by malloc.asm).
this required splitting the forced size argument of the binary op format into an R/M and an IMM forced size.
this should have no performance penalty, as we examine exactly one of them on each execution.
using this, i've changed the shift and shiftx instructions to have their standard operand sizes while still being spatially efficient for imm operands.

the c++ impl for the btx family was apparently non-standard for memory operands, as it just performed an in-place modulo bit logic.
the correct behavior (as is in this commit) treats a memory op as a bit string extending as far as needed into memory.