malloc tests
added tests for the core malloc functions (malloc and free).
realloc is currently not tested and calloc is currently not supported (just commented out).

the impl for malloc/free was brought over from the csx64-stdlib repository, but it didn't have automated testing.
testing revealed that free() wasn't updating the 'prev' pointer in the linked list merge operations.
this was likely from the first impl, before it was a doubly-linked list.
the tests for malloc/free now invasively explore the entire malloc structure from the rust interface to ensure it is properly updated.

added the DEBUG_CPU instruction (output not finalized) to help find the free() bug.

free(null) is now a no-op (previously was undefined behavior).