realloc tests + malloc changes
added the tests for realloc.
this found a number of problems in the previous impl which weren't caught before due to a lack of usage.

made a few changes to the way malloc works.
we no longer require buckets to be non-empty (zero size allocation block).
this enables some more efficient merging logic in arbitrary free sequences.
additionally, we now require that the trailing block in the malloc chain be occupied, with vacant tails being pruned from the structure.
this will make it easier to implement a trimming operation.
but more importantly, it means there are a few places where we no longer need to do bounds checking for dereferences.
(these places were in realloc(), and apparently weren't being checked, so it's a good thing we no longer need to).