malloc/realloc bug fix
i added one last test for the malloc/realloc logic to represent average use case.
it allocates a bunch of buckets of random sizes with random content, then in random order resizes them randomly, and fills with new random content.
we check that contents are preserved with realloc, that allocations are in proper absolute positions, and that there's no overlapping.

this occasionally caught a mem out of bounds failure in the malloc/realloc logic.
by setting the __malloc_step constant to 1 (instead of 1MB), i was able to find the problems and fix them.
the issue was that we would sometimes dereference past __malloc_end, which is not guaranteed to be within the program break established by sys_brk.
the grace block of 1MB was hiding the problem in most executions, including all previous tests.

they should now be completely trustworthy.

now that the problem is fixed, changed the step size back to 1MB.
given how we store memory in the emulator, i'm uncertain that allocating pages is actually more efficient
than just doing a sys_brk every time we need to increase memory (benching would be needed for that).
however, it is more realistic in real-world systems with virtual address spaces, so i'll leave it for now.