added exit, atexit, and abort
added exit/atexit and abort to the stdlib files (with tests).
the impls for exit and atexit are different than the in the cpp stdlib version:
apparently the minimum supported number of atexit hooks is 32.
we used to allow an arbitrary number, but they were allocated dynamically.
however, if the system memory is very small it might not support the required 32.
thus, for now, i've changed it over to just be a static array of size 32.
later, we can decide to also support dynamic arrays if the static is filled (but dynamic alone is not ok).

changed value op encoding's default_imm_size to default_size.
only the jump instructions were using it, and i'd like them to default to 64-bit memory values as they do for imm.
(this came up in exit() impl).

replaced several usages of std::io::Cursor with plain Vec (mostly just in test functions).