$bin update + expr parse bug fix
$bin used to be a critical expression, much like the declare directive family.
this was because it was handled during expr parsing, which only happens once.
i've moved that part of the logic into a meta expr op, and moved the binary string set into the symbol table.
this means we can defer string interning until later - currently, any time during assemble time is acceptable.
this can be further extended to allow some cases to slip into link time, but not all cases are safe to do this.
conservatively, this commit prohibits all link time interning.

fixed a bug in the expr parser where it was using the wrong index offset for nested expression parsing inside macros.

updated all the crates we've been using to latest releases.