linker overhaul
massive changes to the linker.
reverted half of the last commit (where we moved binary interning into Expr evaluation).
that structure is still used, but the expr evaluator always fails to evaluate it.
it's now entirely on the linker to process the $bin requests by directly manipulating the expression trees.

in a dream i envisioned a utopia where all the files could be merged together, so i did that.
we now mutate all personal symbols to be unique e.g. 'foo' -> 'foo:0'.
this way, we can merge all the symbols into one table (and also merge the separate segment holes into one per segment).
this lets us get away with a lot of things we couldn't do before.
for instance, we used to require globals be evaluated before being imported, meaning you couldn't define a global in terms of an extern.
that's no longer the case, and you can do literally whatever you want so long as it would be valid when written as a single file.

minor addition: added overloads of '+' for Binary+Character and Character+Binary, which were missing.